# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OfferOps API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /todo-tasks:
    get:
      operationId: listTodoTasks
      tags: [todo-tasks]
      summary: List todo tasks
      parameters:
        - name: employee_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: task_type
          in: query
          schema:
            type: string
            enum: [CREATE_IOS_TRACKER_CAMPAIGN, CREATE_ANDROID_TRACKER_CAMPAIGN, FIND_WINNERS, PAUSE_TRAFFIC_SOURCE_CAMPAIGNS]
        - name: workspace_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of todo tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TodoTask"

    post:
      operationId: createTodoTask
      tags: [todo-tasks]
      summary: Create todo task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoTaskBody"
      responses:
        "201":
          description: Todo task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoTask"
  /todo-tasks/{id}:
    get:
      operationId: getTodoTask
      tags: [todo-tasks]
      summary: Get todo task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Todo task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoTask"

    patch:
      operationId: updateTodoTask
      tags: [todo-tasks]
      summary: Update todo task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTodoTaskBody"
      responses:
        "200":
          description: Updated todo task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoTask"

    delete:
      operationId: deleteTodoTask
      tags: [todo-tasks]
      summary: Delete todo task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    TodoTask:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        relatedBatchId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        taskType:
          type: string
          enum: [CREATE_IOS_TRACKER_CAMPAIGN, CREATE_ANDROID_TRACKER_CAMPAIGN, GO_LIVE_TRAFFIC_SOURCE_CAMPAIGN, MOVE_WINNERS_TO_SCALED_CAMPAIGN, FIND_WINNERS, PAUSE_TRAFFIC_SOURCE_CAMPAIGNS, CREATE_IOS_CAMPAIGN, CREATE_ANDROID_CAMPAIGN, GO_LIVE, OPTIMIZATION_FOLLOWUP]
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [TODO, IN_PROGRESS, BLOCKED, DONE]
        dueDate:
          type: ["string", "null"]
        workspaceId:
          type: ["integer", "null"]
        trafficSourceName:
          type: ["string", "null"]
        device:
          type: ["string", "null"]
        createdAt:
          type: string
        employeeName:
          type: ["string", "null"]
        batchName:
          type: ["string", "null"]
      required:
        - id
        - employeeId
        - title
        - taskType
        - priority
        - status
        - createdAt
    CreateTodoTaskBody:
      type: object
      properties:
        employeeId:
          type: integer
        relatedBatchId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        taskType:
          type: string
          enum: [CREATE_IOS_TRACKER_CAMPAIGN, CREATE_ANDROID_TRACKER_CAMPAIGN, GO_LIVE_TRAFFIC_SOURCE_CAMPAIGN, MOVE_WINNERS_TO_SCALED_CAMPAIGN, FIND_WINNERS, PAUSE_TRAFFIC_SOURCE_CAMPAIGNS, CREATE_IOS_CAMPAIGN, CREATE_ANDROID_CAMPAIGN, GO_LIVE, OPTIMIZATION_FOLLOWUP]
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [TODO, IN_PROGRESS, BLOCKED, DONE]
        dueDate:
          type: ["string", "null"]
        workspaceId:
          type: ["integer", "null"]
        trafficSourceName:
          type: ["string", "null"]
        device:
          type: ["string", "null"]
      required:
        - employeeId
        - title
        - taskType
        - priority
    UpdateTodoTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        taskType:
          type: string
          enum: [CREATE_IOS_TRACKER_CAMPAIGN, CREATE_ANDROID_TRACKER_CAMPAIGN, GO_LIVE_TRAFFIC_SOURCE_CAMPAIGN, MOVE_WINNERS_TO_SCALED_CAMPAIGN, FIND_WINNERS, PAUSE_TRAFFIC_SOURCE_CAMPAIGNS, CREATE_IOS_CAMPAIGN, CREATE_ANDROID_CAMPAIGN, GO_LIVE, OPTIMIZATION_FOLLOWUP]
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [TODO, IN_PROGRESS, BLOCKED, DONE]
        dueDate:
          type: ["string", "null"]
        relatedBatchId:
          type: ["integer", "null"]
        blockedReason:
          type: ["string", "null"]
...